requirements.txt
fastapi
uvicorn
diffusers
torch
pillow
invisible_watermark
transformers
accelerate
safetensors